Combined all custom datatypes into a single file, removed unnecessary usings, added more clarity to the naming of pubsub functions
