+++ samples/CosmosDBIntegration/WriteAroundSamples/StreamSample.cs
@@ -1,10 +1,8 @@
-﻿using Microsoft.Azure.WebJobs;
+﻿using Microsoft.Azure.WebJobs.Extensions.Redis.Samples.Models;
 using Microsoft.Extensions.Logging;
 using StackExchange.Redis;
 using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 
 namespace Microsoft.Azure.WebJobs.Extensions.Redis.Samples
